# Tweaks and more LoopVectorization support