--- a/Project.toml
+++ b/Project.toml
@@ -10,9 +10,9 @@ VectorizationBase = "3d5dd08c-fd9d-11e8-17fa-ed2836048c2f"
 
 [compat]
 julia = "1"
-LoopVectorization = "0.12"
+LoopVectorization = "0.12.3"
 TropicalNumbers = "0.2"
-VectorizationBase = "0.19"
+VectorizationBase = "0.19.10"
 
 [extras]
 Octavian = "6fd5a793-0b7e-452c-907f-f8bfe9c57db4"