Also some fixes, like `zero_vecunroll` should return a bunch of `Vec`s of `-Inf`.

I also really don't like those promote definitions.
When where they needed? Do you have an example?
I suspect they aren't, but didn't want to touch them without some test case I could use to confirm.

I am also use `Base.FastMath.max_fast` instead of `max`, as this cuts out a few instructions. It didn't have much of an impact on performance, but it cleans up the assembly nicely.

Benchmarks on my computer:
```julia
julia> A = randn(1000,1000);

julia> B = randn(1000,1000);

julia> Cref = A*B; C = similar(Cref);

julia> At = Tropical.(A);

julia> Bt = Tropical.(B);

julia> Ctref = At * Bt; Ct = similar(Ctref);

julia> @benchmark Octavian.matmul_serial!($C,$A,$B)
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     17.310 ms (0.00% GC)
  median time:      17.409 ms (0.00% GC)
  mean time:        17.409 ms (0.00% GC)
  maximum time:     18.099 ms (0.00% GC)
  --------------
  samples:          288
  evals/sample:     1

julia> 2e-9*1000^3 / 17.31e-3
115.54015020219526

julia> C ≈ Cref
true

julia> @benchmark Octavian.matmul_serial!($Ct,$At,$Bt)
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     31.906 ms (0.00% GC)
  median time:      31.946 ms (0.00% GC)
  mean time:        31.960 ms (0.00% GC)
  maximum time:     32.764 ms (0.00% GC)
  --------------
  samples:          157
  evals/sample:     1

julia> 2e-9*1000^3 / 31.906e-3
62.684134645521226

julia> Ct ≈ Ctref
true

julia> 4.1 * 2 * 16 # peak theoretical `Float64` gflops
131.2

julia> 4.1 * 2 * 8 # peak theoretical `Tropical{Float64}` gflops
65.6
```
We get `62.7` out of the theoretical peak of `65.6` gflops on my machine.
I think that's pretty good; very little performance left on the table.

Peak GFLOPS is calculated as clock speed * instructions/clock * ops/instruction.
`Float64` has twice the theoretical peak thanks to fused multiply add instructions, i.e. 16 `Float64` operations per fma with AVX512, while for `Tropical{Float64}` it has 8 operations per `max` and per `+`.