@GiggleLiu 2021-03-19
Also, I need the following patch to work in my local host. I can not figure out why my host is so different.

https://github.com/TensorBFS/TropicalGEMM.jl/pull/2

FYI:
```
(@v1.6) pkg> st
      Status `~/.julia/environments/v1.6/Project.toml`
  [1520ce14] AbstractTrees v0.3.4
  [a9ab73d0] BatchedRoutines v0.2.1
  [6e4b80f9] BenchmarkTools v0.5.0
  [d7b10767] BinarySparseTensors v0.1.0 `~/.julia/dev/BinarySparseTensors`
  [50ba71b6] BitBasis v0.7.2
  [81fc84e3] BliContractor v1.1.0 `~/.julia/dev/BliContractor`
  [336ed68f] CSV v0.8.4
  [052768ef] CUDA v2.6.2
  [159f3aea] Cairo v1.0.5
  [5ae59095] Colors v0.12.6
  [a81c6b42] Compose v0.9.2
  [150eb455] CoordinateTransformations v0.6.1
  [717857b8] DSP v0.6.10
  [a93c6f00] DataFrames v0.22.5
  [1313f7d8] DataFramesMeta v0.6.0
  [e30172f5] Documenter v0.26.3
  [35a29f4d] DocumenterTools v0.1.9
  [497a8b3b] DoubleFloats v1.1.18
  [b3ff564c] EliminateGraphs v0.1.0
  [d4d017d3] ExponentialUtilities v1.8.0 `~/.julia/dev/ExponentialUtilities`
  [7a1cc6ca] FFTW v1.3.2
  [5789e2e9] FileIO v1.6.4
  [652a1917] Fire v0.1.1
  [53c48c17] FixedPointNumbers v0.8.4
  [f6369f11] ForwardDiff v0.10.17
  [01680d73] GenericSVD v0.3.0
  [a2cc645c] GraphPlot v0.4.4
  [708ec375] Gumbo v0.8.0
  [f67ccb44] HDF5 v0.15.4
  [e47d643f] HierarchicalBipartition v0.1.0 `../../dev/HierarchicalBipartition`
  [9136182c] ITensors v0.1.40
  [42fd0dbc] IterativeSolvers v0.9.0
  [033835bb] JLD2 v0.2.4
  [682c06a0] JSON v0.21.1
  [e5e0dc1b] Juno v0.8.4
  [63c18a36] KernelAbstractions v0.5.3
  [b964fa9f] LaTeXStrings v1.2.1
  [23fbe1c1] Latexify v0.14.11
  [ac33ec98] LightBayesian v0.2.3 `~/.julia/dev/LightBayesian`
  [093fc24a] LightGraphs v1.3.5
  [98b081ad] Literate v2.8.0
  [aa2f6b4e] LogarithmicNumbers v0.4.0
  [bdcacae8] LoopVectorization v0.12.2
  [ae8d54c2] Luxor v2.10.0
  [0fe46d8b] MISExperimentUtils v0.5.6 `~/.julia/dev/MISExperimentUtils`
  [33e6dc65] MKL v0.4.0 `~/.julia/dev/MKL`
  [5dd3f0b1] MatchCore v0.1.0
  [ab4ef3a6] NiLang v0.8.2 `~/.julia/dev/NiLang`
  [575d3204] NiLangCore v0.8.2 `~/.julia/dev/NiLangCore`
  [ebe7aa44] OMEinsum v0.3.3 `~/.julia/dev/OMEinsum`
  [6fd5a793] Octavian v0.2.11
  [429524aa] Optim v1.2.4
  [32113eaa] PkgBenchmark v0.2.10
  [14b8a8f1] PkgTemplates v0.7.16
  [91a5bcdd] Plots v1.10.6
  [c3e4b0f8] Pluto v0.12.21
  [ef6358c6] PlutoMustache v0.2.4 `~/.julia/dev/PlutoMustache`
  [7f904dfe] PlutoUI v0.7.1
  [a5ec66cb] PlutoUtils v0.1.0 `https://github.com/GiggleLiu/PlutoUtils.jl#static-export`
  [ade5400d] Pomicon v0.3.4 `~/.julia/dev/Pomicon`
  [c46f51b8] ProfileView v0.6.9
  [438e738f] PyCall v1.92.2
  [d330b81b] PyPlot v2.9.0
  [ae029012] Requires v1.1.3
  [37e2e3b7] ReverseDiff v1.7.0
  [d4ee886f] ReversibleSeismic v0.1.0 `../../dev/ReversibleSeismic`
  [295af30f] Revise v3.1.14
  [6038ab10] Rotations v1.0.2
  [476501e8] SLEEFPirates v0.6.12
  [4456351a] SimpleTensorNetworks v0.2.1 `~/.julia/dev/SimpleTensorNetworks`
  [aa65fe97] SnoopCompile v2.6.0
  [90137ffa] StaticArrays v0.12.5
  [f3b207a7] StatsPlots v0.14.19
  [6d0aa2be] StochasticOptimizers v0.8.2 `~/.julia/dev/StochasticOptimizers`
  [123dc426] SymEngine v0.8.3
  [6aa20fa7] TensorOperations v3.1.0
  [8290d209] ThreadingUtilities v0.4.1
  [a4ad3063] TropicalGEMM v0.1.0 `../../dev/TropicalGEMM`
  [30ce92b6] TropicalMIS v0.1.2 `../../dev/TropicalMIS`
  [b3a74e9c] TropicalNumbers v0.2.2
  [d36a0d72] TropicalTensors v0.2.0 `~/.julia/dev/TropicalTensors`
  [9d95972d] TupleTools v1.2.0
  [1986cc42] Unitful v1.6.0
  [3d5dd08c] VectorizationBase v0.19.9
  [52a3aca4] Viznet v0.3.2 `~/.julia/dev/Viznet`
  [5872b779] Yao v0.6.3
  [32cfe2d9] YaoPlots v0.6.0 `~/.julia/dev/YaoPlots`
```